+++ oshdb-util/src/test/java/org/heigit/ohsome/oshdb/util/geometry/OSHDBGeometryBuilderTest.java
@@ -6,9 +6,13 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import java.util.function.Consumer;
 import org.heigit.ohsome.oshdb.OSHDBBoundingBox;
 import org.heigit.ohsome.oshdb.OSHDBTimestamp;
+import org.heigit.ohsome.oshdb.osm.OSM;
 import org.heigit.ohsome.oshdb.osm.OSMEntity;
+import org.heigit.ohsome.oshdb.osm.OSMMember;
+import org.heigit.ohsome.oshdb.osm.OSMRelation;
 import org.heigit.ohsome.oshdb.util.geometry.helpers.FakeTagInterpreterAreaAlways;
 import org.heigit.ohsome.oshdb.util.geometry.helpers.FakeTagInterpreterAreaMultipolygonAllOuters;
 import org.heigit.ohsome.oshdb.util.geometry.helpers.FakeTagInterpreterAreaNever;
@@ -18,6 +22,7 @@ import org.junit.jupiter.api.Test;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
+import org.locationtech.jts.geom.MultiPolygon;
 import org.locationtech.jts.geom.Point;
 import org.locationtech.jts.geom.Polygon;
 import org.locationtech.jts.io.ParseException;
@@ -29,8 +34,11 @@ import org.locationtech.jts.io.WKTReader;
 class OSHDBGeometryBuilderTest extends OSHDBGeometryTest {
   private static final double DELTA = 1E-6;
 
-  OSHDBGeometryBuilderTest() {
-    super("./src/test/resources/geometryBuilder.osh");
+  public OSHDBGeometryBuilderTest() {
+    super(
+        "./src/test/resources/geometryBuilder.osh",
+        "./src/test/resources/relations/multipolygonShellsShareNode.osm"
+    );
   }
 
   @Test
@@ -208,5 +216,112 @@ class OSHDBGeometryBuilderTest extends OSHDBGeometryTest {
       new Coordinate(0, 1),
       new Coordinate(0, 0)};
     assertArrayEquals(test, geometry.getCoordinates());
+  }
+
+  /** Test building of multipolygons with self-touching outer rings.
+   *
+   * <p>Example:
+   * <pre>
+   * lat
+   *  ^
+   *  |
+   *  4 a--d
+   *  3 |  |
+   *  2 b--c--e
+   *  1    |  |
+   *  0    g--f
+   *    0 1 2 3456 -> lon
+   *
+   * A: (a,b,c)\
+   *            one shell (a,b,c,d,a)
+   * B: (c,d,a)/
+   *
+   * C: (c,e,f)\
+   *            one shell (c,e,f,g,c)
+   * D: (f,g,c)/
+   * </pre></p>
+   *
+   * <p>Expected:
+   * <pre>
+   * - MULTIPOLYGON (((0 4, 0 2, 3 2, 3 4, 0 4)), ((3 2, 6 2, 6 0, 3 0, 3 2)))
+   * - MULTIPOLYGON (((3 2, 6 2, 6 0, 3 0, 3 2)), ((0 4, 0 2, 3 2, 3 4, 0 4)))
+   * - or similar
+   * </pre></p>
+   */
+  @Test
+  public void testMultipolygonShellsShareNode() {
+    // single figure-8
+    var members = testData.relations().get(100L).get(0).getMembers();
+    getMultipolygonSharedNodeCheck(geom -> {
+      assertTrue(geom instanceof MultiPolygon);
+      assertEquals(2, geom.getNumGeometries());
+    }).accept(members);
+    // all permutations of relation members should lead to the same result
+    members = testData.relations().get(101L).get(0).getMembers();
+    checkAllMemberPermutations(members.length, members, getMultipolygonSharedNodeCheck(geom -> {
+      assertTrue(geom instanceof MultiPolygon);
+      assertEquals(2, geom.getNumGeometries());
+    }));
+    // double loop
+    members = testData.relations().get(102L).get(0).getMembers();
+    getMultipolygonSharedNodeCheck(geom -> {
+      assertTrue(geom instanceof MultiPolygon);
+      assertEquals(3, geom.getNumGeometries());
+    }).accept(members);
+    // second loop forming whole
+    members = testData.relations().get(103L).get(0).getMembers();
+    getMultipolygonSharedNodeCheck(geom -> {
+      assertTrue(geom instanceof Polygon);
+      assertEquals(1, ((Polygon) geom).getNumInteriorRing());
+    }).accept(members);
+    // multiple holes
+    members = testData.relations().get(104L).get(0).getMembers();
+    getMultipolygonSharedNodeCheck(geom -> {
+      assertTrue(geom instanceof Polygon);
+      assertEquals(2, ((Polygon) geom).getNumInteriorRing());
+    }).accept(members);
+    // hole in hole = additional outer
+    members = testData.relations().get(105L).get(0).getMembers();
+    checkAllMemberPermutations(members.length, members, getMultipolygonSharedNodeCheck(geom -> {
+      assertTrue(geom instanceof MultiPolygon);
+      assertEquals(2, geom.getNumGeometries());
+      assertEquals(1,
+          ((Polygon) geom.getGeometryN(0)).getNumInteriorRing()
+              + ((Polygon) geom.getGeometryN(1)).getNumInteriorRing()
+      );
+    }));
+  }
+
+  private Consumer<OSMMember[]> getMultipolygonSharedNodeCheck(Consumer<Geometry> tester) {
+    var areaDecider = new FakeTagInterpreterAreaMultipolygonAllOuters();
+    var timestamp = TimestampParser.toOSHDBTimestamp("2001-01-01");
+    return relMembers -> {
+      var relation = OSM.relation(1, 1, timestamp.getEpochSecond(), 0, 0, null, relMembers);
+      var geom = OSHDBGeometryBuilder.getGeometry(relation, timestamp, areaDecider);
+      assertTrue(geom.isValid());
+      tester.accept(geom);
+    };
+  }
+
+  private static <T> void checkAllMemberPermutations(int n, T[] elements, Consumer<T[]> consumer) {
+    if (n == 1) {
+      consumer.accept(elements);
+    } else {
+      for (int i = 0; i < n - 1; i++) {
+        checkAllMemberPermutations(n - 1, elements, consumer);
+        if (n % 2 == 0) {
+          swap(elements, i, n - 1);
+        } else {
+          swap(elements, 0, n - 1);
+        }
+      }
+      checkAllMemberPermutations(n - 1, elements, consumer);
+    }
+  }
+
+  private static <T> void swap(T[] elements, int a, int b) {
+    T tmp = elements[a];
+    elements[a] = elements[b];
+    elements[b] = tmp;
   }
 }